+++ src/meta-esdt/issue-meta-esdt.ts
@@ -66,8 +66,9 @@ const promptQuestions: PromptObject[] = [
     name: 'tokenProperties',
     message: 'Please choose token properties.\n',
     choices: sftNftTokenProperties.map((property) => ({
-      title: property,
-      value: property,
+      title: property.name,
+      value: property.name,
+      description: property.description,
     })),
   },
 ];
@@ -100,11 +101,11 @@ export const issueMetaEsdt = async () => {
     for (const property of sftNftTokenProperties) {
       let propertyEnabled = false;
 
-      if (tokenProperties.includes(property)) {
+      if (tokenProperties.includes(property.name)) {
         propertyEnabled = true;
       }
 
-      args.push(BytesValue.fromUTF8(property));
+      args.push(BytesValue.fromUTF8(property.name));
       args.push(BytesValue.fromUTF8(propertyEnabled.toString()));
     }