+++ src/sft/toggle-special-roles-sft.ts
@@ -43,8 +43,9 @@ const promptQuestions = (type: OperationType): PromptObject[] => [
       type === 'set' ? 'assign' : 'remove'
     }.\n`,
     choices: sftTokenSpecialRoles.map((property) => ({
-      title: property,
-      value: property,
+      title: property.name,
+      value: property.name,
+      description: property.description,
     })),
   },
 ];
@@ -70,8 +71,8 @@ export const toggleSpecialRolesSft = async (type: OperationType) => {
     ];
 
     for (const role of sftTokenSpecialRoles) {
-      if (specialRoles.includes(role)) {
-        args.push(BytesValue.fromUTF8(role));
+      if (specialRoles.includes(role.name)) {
+        args.push(BytesValue.fromUTF8(role.name));
       }
     }